requirements.txt
requests==2.32.3
PyYAML>=6.0